# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: K&S Solar Energy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /technicians:
    get:
      operationId: getTechnicians
      tags: [technicians]
      summary: Get all technicians (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin only
    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create a new technician account (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnicianRequest"
      responses:
        "201":
          description: Technician created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "400":
          description: Validation error or email already exists
        "403":
          description: Forbidden
  /technicians/{id}:
    patch:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update technician info or deactivate (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTechnicianRequest"
      responses:
        "200":
          description: Technician updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /warranties/search:
    get:
      operationId: searchWarrantyByInvoice
      tags: [warranties]
      summary: Search warranty by invoice number (public, no auth required)
      parameters:
        - name: invoiceNumber
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warranty found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "404":
          description: Not found
  /warranties:
    get:
      operationId: getWarranties
      tags: [warranties]
      summary: Get warranties (own or filtered by userId for admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of warranties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warranty"
        "401":
          description: Unauthorized
    post:
      operationId: createWarranty
      tags: [warranties]
      summary: Add a warranty record for a customer (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarrantyRequest"
      responses:
        "201":
          description: Warranty created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "400":
          description: Validation error
        "403":
          description: Forbidden
  /warranties/{id}:
    patch:
      operationId: updateWarranty
      tags: [warranties]
      summary: Update a warranty record (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWarrantyRequest"
      responses:
        "200":
          description: Warranty updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteWarranty
      tags: [warranties]
      summary: Delete a warranty record (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warranty deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Technician:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialty:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - status
        - createdAt
    CreateTechnicianRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
        specialty:
          type: string
      required:
        - name
        - email
        - phone
        - password
    UpdateTechnicianRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialty:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
...
    Warranty:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        warrantyType:
          type: string
          enum: [inverter, panels, battery, installation]
        brand:
          type: string
        model:
          type: string
          nullable: true
        purchaseDate:
          type: string
        durationMonths:
          type: integer
        notes:
          type: string
          nullable: true
        expiryDate:
          type: string
        warrantyStatus:
          type: string
          enum: [active, expiring_soon, expired]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - warrantyType
        - brand
        - purchaseDate
        - durationMonths
        - expiryDate
        - warrantyStatus
        - createdAt
    CreateWarrantyRequest:
      type: object
      properties:
        userId:
          type: string
        invoiceNumber:
          type: string
        warrantyType:
          type: string
          enum: [inverter, panels, battery, installation]
        brand:
          type: string
        model:
          type: string
        purchaseDate:
          type: string
        durationMonths:
          type: integer
          minimum: 1
        notes:
          type: string
      required:
        - userId
        - warrantyType
        - brand
        - purchaseDate
        - durationMonths
    UpdateWarrantyRequest:
      type: object
      properties:
        invoiceNumber:
          type: string
          nullable: true
        warrantyType:
          type: string
          enum: [inverter, panels, battery, installation]
        brand:
          type: string
        model:
          type: string
          nullable: true
        purchaseDate:
          type: string
        durationMonths:
          type: integer
          minimum: 1
        notes:
          type: string
          nullable: true
...